Improving look & feel of iOS 7 bubbles
I adjusted the bubble-min images to have corner radii more closely
matching iOS 7’s, and attempted to get them a bit sharper looking.
Adjustments made to insets to sharpen up the appearance of the bubble
on the top/right edges. Adjusted x and y position of the textview frame
to improve padding and position text more closely to vertical center
(this part’s kind of a hack… probably a better way than just
hard-coding +4 and +3!)