Fix scroll to bottom issue if layout changes view size.
If, for example on the iPhone 5, the initial layout enlarges the view, then the initial scroll-to-bottom call leaves the table view scrolled to something that is not quite the end (offset by the change in size). On the initial layout only we should trigger the scroll-to-bottom rather than in the viewWillAppear callback.