There are a lot of changes in several files, but I felt I should keep them all in one commit since they're all directly related to incorporating measuring the size of attributedText rather than just text.  This is in order to have multi-line notifications and proper sizing.
One of the biggest changes is adding the method attributedTextForRowAtIndexPath: to JSMessagesViewDelegate.  This is necessary as it needs to compute the height for the row (UITableViewDelegate).  Also the cell will size itself depending on whether or not attributedText is passed into neededHeightForBubbleMessageCellWithText:.

Also included is a couple of computations fix up the bubbles.