Introduce property resignFirstResponderOnKeyboardDidHide
Pull Request #646 introduced the callback keyboardControllerKeyboardDidHide to set the toolbar bottom layout guide constant to 0 to resolve a specific issue.
This has the side effect of causing issue #763.

This commit introduces a switch that allows the consumer to choose which behavior they need.
More information can be seen in the comment added in JSQMessagesViewController