Don't return spring behavior for attributes with empty size.
The demo project currently displays a new outgoing message using JSQMessagesViewController -finishSendingMessageAnimated:, which calls UICollectionView -reloadData. This works fine. However when new items are instead added using UICollectionView -insertItemsAtIndexPaths:, a failure occurs in the flow layout when trying to add the spring behavior to the UIDynamicAnimator in -prepareForCollectionViewUpdates:. This change protects against that failure, which is caused by an invalid zero size in the layout attributes. Adding a nil behavior instead behaves correctly.