move default dataDetector setup into textView awakeFromNib
This makes sure that we are not enabling dataDetectors in every
cellAtIndexPath method call even if the library user wants to disable
it.

When creating it we set the default setting, then the user can do
whatever he wants with it.

Also there’s no reason to reset the dataDetectors to None on reuse, as
the cell probably wants to use the previous settings. If not the
library user should explicitly change it anyway.