Fix for issue 1174
This avoids the problems mentioned in #1174 by removing our target on the Interactive Pop Gesture at viewDidDisappear-time
It avoid regressing to the problems documented in #257 by capturing the UIGestureRecognizer (as weak) at the time we addTarget. This allow us to remember on whom we should removeTarget.

I will have a separate commit to add a 2-level navigation-controller push to the Demo project, but that is not directly related to #1174 nor #257