Bug Fix #1449 - No opportunity to change JSQMessagesInputToolbar preferredDefaultHeight.
It doesn't seem like you can change the height of the inputToolbar using the documented property preferredDefaultHeight since it is used to configure the height constraint in viewDidLoad.
In order to fix the bug I added the line of code:

self.toolbarHeightConstraint.constant = self.inputToolbar.preferredDefaultHeight;

In the method:

- (void)jsq_updateKeyboardTriggerPoint

This function called from viewWillAppear and the view load with the correct constraints.

I'm not really sure if it's the optimal location for this line of code, alternative location is just to call from viewWillAppear.