Allow users to create their own views inside message bubbles.
The JSQCustomMediaItem class is the starting place for allowing users to
create their own custom views. It manages the communication between the
cell/collection view and the custom view.

The JSQCustomMediaView is the foundation for all custom views. Any
custom view either needs to use an instance of this class and add their
own viewes to it or subclass it.

The JSQHTMLMessageView is a subclass of JSQCustomMediaView and allows
HTML to be displayed inside a message bubble.

Basic unit tests as well as enhancements to the demo are included.